requirements/dev.txt
42cc-pystyle==0.0.17
black==22.3.0
coverage==6.3.3
dj-database-url==1.0.0
Django==4.0.4
django-cors-headers==3.13.0
django-oauth-toolkit==2.1.0
django-webpack4-loader==0.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt
drf-social-oauth2==1.2.1
flake8==3.8.3
flake8-absolute-import==1.0.0.1
flake8-black==0.3.3
flake8-broken-line==0.4.0
flake8-cognitive-complexity==0.1.0
flake8-commas==2.1.0
flake8-django==1.1.1
flake8-eradicate==1.2.1
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-print==5.0.0
flake8-variables-names==0.0.5
gunicorn==20.1.0
oauthlib==3.2.1
pathspec==0.10.1
PyJWT
python-dotenv==0.20.0


